refactor: 모달 컴포넌트 구조 변경 및 추상화
