+++ modal-components/src/App.tsx
@@ -1,14 +1,14 @@
-import BasicModal from "./lib/components/BasicModal";
+import BasicModal from "./lib/components/modals/BasicModal";
 import "./App.css";
 import {
   useCardNumbersInput,
   useCardCVCInput,
 } from "../../payments-hooks/src/lib";
 import { useState } from "react";
-import ConfirmModal from "./lib/components/ConfirmModal";
-import AlertModal from "./lib/components/AlertModal";
-import PromptModal from "./lib/components/PromptModal";
-import AgreementModal from "./lib/components/AgreementModal";
+import ConfirmModal from "./lib/components/modals/ConfirmModal";
+import AlertModal from "./lib/components/modals/AlertModal";
+import PromptModal from "./lib/components/modals/PromptModal";
+import AgreementModal from "./lib/components/modals/AgreementModal";
 
 function CardNumberInput() {
   const { cardNumbersInfo, cardBrand, cardBlocks, onChangeHandler } =
