+++ modal-components/src/lib/components/modals/PromptModal.tsx
@@ -1,0 +1,57 @@
+import ModalLayout from "./common/ModalLayout";
+import Button from "../common/Button";
+import Input from "../common/Input";
+import styled from "@emotion/styled";
+import { ConfirmModalProps } from "../../types/modalTypes";
+
+const ButtonContainer = styled.div`
+  width: 100%;
+  height: 36px;
+  display: flex;
+  justify-content: flex-end;
+  gap: 8px;
+`;
+
+const PromptModal = ({
+  modalPosition,
+  modalSize,
+  titleText,
+  descriptionText,
+  closeType,
+  onClose,
+  onConfirm = onClose,
+  onCancel = onClose,
+}: ConfirmModalProps) => {
+  const footer = (
+    <ButtonContainer>
+      <Button widthSize="80px" onClick={onCancel} variant="secondary">
+        취소
+      </Button>
+      <Button widthSize="80px" onClick={onConfirm} variant="primary">
+        확인
+      </Button>
+    </ButtonContainer>
+  );
+
+  const customForm = (
+    <form>
+      <Input />
+    </form>
+  );
+
+  return (
+    <ModalLayout
+      modalPosition={modalPosition}
+      modalSize={modalSize}
+      titleText={titleText}
+      descriptionText={descriptionText}
+      closeType={closeType}
+      onClose={onClose}
+      footer={footer}
+    >
+      {customForm}
+    </ModalLayout>
+  );
+};
+
+export default PromptModal;
