+++ modal-components/src/lib/index.ts
@@ -1,4 +1,4 @@
 export { useModal } from "./hook/useModal";
-export { default as BasicModal } from "./components/BasicModal";
-export * from "./styles/ModalStyle";
+export { default as BasicModal } from "./components/modals/BasicModal";
+export * from "./components/modals/common/ModalStyle";
 export * from "./styles/ModalTextStyle";
